incoming/zipparu/pyproject.toml
[project]
name = "zipparu"
version = "0.1.0"
description = "Zip and upload a folder filtered by .gitignore"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "gitignore_parser",
    "requests"
]

[project.scripts]
zipparu = "zipparu.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"